pisap: remove duplicates.
