--- pisap/numerics/__init__.py
@@ -1,15 +1,0 @@
-# -*- coding: utf-8 -*-
-##########################################################################
-# XXX - Copyright (C) XXX, 2017
-# Distributed under the terms of the CeCILL-B license, as published by
-# the CEA-CNRS-INRIA. Refer to the LICENSE file or to
-# http://www.cecill.info/licences/Licence_CeCILL-B_V1-en.html
-# for details.
-##########################################################################
-
-"""
-This module defines common optimization strategies with predefined operators,
-gradients, costs, ...
-"""
-from .reconstruct import sparse_rec_condat_vu
-from .reconstruct import sparse_rec_fista
