--- pisap/stats/__init__.py
@@ -1,17 +1,0 @@
-# -*- coding: utf-8 -*-
-##########################################################################
-# XXX - Copyright (C) XXX, 2017
-# Distributed under the terms of the CeCILL-B license, as published by
-# the CEA-CNRS-INRIA. Refer to the LICENSE file or to
-# http://www.cecill.info/licences/Licence_CeCILL-B_V1-en.html
-# for details.
-##########################################################################
-
-"""
-This module defines common statistical functions or utility functions.
-"""
-
-
-from .noise import mad
-from .noise import sigma_mad
-from .noise import multiscale_sigma_mad
